pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=42", "wheel"]

[project]
dependencies = [
  "aiohttp>=3.11.13",
  "anthropic>=0.45.2",
  "cachetools>=5.5.2",
  "chromadb>=0.6.3",
  "debugpy>=1.8.12",
  "dotenv>=0.9.9",
  "faiss-cpu>=1.10.0",
  "fitz>=0.0.1.dev2",
  "frontend>=0.0.3",
  "genanki>=0.13.1",
  "huggingface-hub>=0.28.1",
  "ipython>=9.0.0",
  "langchain-anthropic>=0.3.7",
  "langchain-chroma>=0.2.2",
  "langchain-community>=0.3.17",
  "langchain-core>=0.3.35",
  "langchain-experimental>=0.3.4",
  "langchain-text-splitters>=0.3.6",
  "langchain-voyageai>=0.1.4",
  "langchain>=0.3.18",
  "langgraph-prebuilt>=0.1.1",
  "langgraph>=0.3.0",
  "loadenv>=0.1.1",
  "loguru>=0.7.3",
  "open-clip-torch>=2.31.0",
  "openai>=1.62.0",
  "pdf2image>=1.17.0",
  "pi-heif>=0.21.0",
  "pillow>=11.1.0",
  "pre-commit>=4.1.0",
  "pydantic>=2.10.6",
  "pymupdf>=1.25.3",
  "pypdf>=5.3.0",
  "pytesseract",
  "python-dotenv>=1.0.1",
  "pytz>=2025.1",
  "pyyaml>=6.0.2",
  "rapidocr-onnxruntime>=1.4.4",
  "rich>=13.9.4",
  "setuptools>=75.8.2",
  "sympy>=1.13.1",
  "tiktoken>=0.8.0",
  "torch>=2.6.0",
  "transformers>=4.49.0",
  "unstructured-client>=0.30.6",
  "unstructured-inference>=0.8.7",
  "unstructured-pytesseract>=0.3.13",
  "unstructured>=0.16.20",
  "uuid>=1.30"
]
description = "Add your description here"
name = "quiz_me"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[tool.interrogate]
exclude = ["tests", "__init__.py"]
fail-under = 90
ignore-init-method = true
ignore-init-module = true
ignore-regex = []
verbose = 1
whitelist-regex = []

[tool.ruff]
target-version = "py312"

[tool.uv]
dev-dependencies = ["pylint>=2.16.2", "pytest>=7.4.0"]

# add pip installed libraries
